Add install script for professional theme
This installation script is meant to be run from within an instance of
CE that has already been set up.  The purpose of this script is to set
up the default theme.  This includes setting up pages and sections that
are called from the home page.  There are several optional switches that
can be set in the url that runs the script.